Move application root modules to application folder.
