+++ src/app/application/app.component.scss
