+++ src/app/application/app.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 
-import { ApiService } from './services';
+import { ApiService } from '../services';
 
-import '../style/app.scss';
+import '../../style/app.scss';
 
 @Component({
     selector: 'dng-app',
